store: Midway in refactoring
Well, almost done moving a bunch of sub functions that should now be in
store, and also decoupling event writing from publisher, but we need to
fix some tests in pub now.
