+++ sub/sub.go
@@ -5,12 +5,7 @@
 package sub
 
 import (
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"path/filepath"
-	"reflect"
-	"sort"
 
 	"github.com/rjeczalik/notify"
 	"github.com/vancluever/fspubsub/store"
@@ -23,62 +18,6 @@ import (
 // filesystem notifier does not block sending events, so the subscriber will
 // miss events if there is an overrun.
 const defaultBufferSize = 10
-
-// Event represents a single event.
-type Event struct {
-	// The ID of the event. This normally translates to the file name from the
-	// store.
-	ID string
-
-	// The event data.
-	Data interface{}
-}
-
-// eventSlice represents multiple events and implements sort.Interface so that
-// the slice can be sorted by the event struct's implementation of
-// IndexedEvent, if it exists. Data sorted by this function gets returned as
-// just the generic event slice.
-type eventSlice []Event
-
-// Len returns the length of the slice, and helps implement sort.Interface.
-func (p eventSlice) Len() int {
-	return len(p)
-}
-
-// Less takes the value at the index of j, and passes it to the Less method
-// of the event data at the index of i. The function panics if p[j] does not
-// implement IndexedEvent.
-func (p eventSlice) Less(i, j int) bool {
-	v, ok := p[i].Data.(IndexedEvent)
-	if !ok {
-		panic(fmt.Errorf("Event at index %d with type %T does not implement sub.IndexedEvent", i, p[i]))
-	}
-	return v.Less(p[j].Data)
-}
-
-// Swap does a simple swap of i and j, implementing sort.Interface.
-func (p eventSlice) Swap(i, j int) {
-	p[i], p[j] = p[j], p[i]
-}
-
-// IndexedEvent is an interface that implements an event with an index.
-//
-// The index, whatever is chosen in the event, should be able to be sorted with
-// a Less method. This method is similar in purpose to the Less method in the
-// standard library sort package, however, the lower index value commonly
-// denoted as i should be a receiver. A trivial example is below:
-//
-//   type E struct {
-//     Index string
-//   }
-//
-//   func (i E) Less(j interface{}) bool {
-//     return i.Index < j.(E).Index
-//   }
-//
-type IndexedEvent interface {
-	Less(j interface{}) bool
-}
 
 // Subscriber is a simple event subscriber, designed to read events from the
 // file system.
@@ -97,7 +36,7 @@ type Subscriber struct {
 
 	// The internal queue channel. Call Queue to get a valid one-way event
 	// channel.
-	queue chan Event
+	queue chan store.Event
 
 	// The internal completion channel. Call Done to get a valid one-way
 	// completion channel.
@@ -113,7 +52,7 @@ type Subscriber struct {
 
 // Queue returns the event channel. This is buffered to the size of the file
 // system notification buffer.
-func (s *Subscriber) Queue() <-chan Event {
+func (s *Subscriber) Queue() <-chan store.Event {
 	return s.queue
 }
 
@@ -155,7 +94,7 @@ func NewSubscriber(dir string, event interface{}) (*Subscriber, error) {
 
 	s := &Subscriber{
 		Stream: stream,
-		queue:  make(chan Event, defaultBufferSize),
+		queue:  make(chan store.Event, defaultBufferSize),
 		done:   make(chan struct{}, 1),
 		errch:  make(chan error, 1),
 	}
@@ -172,7 +111,7 @@ func (s *Subscriber) watch(c chan notify.EventInfo) {
 	for {
 		select {
 		case ei := <-c:
-			e, err := decodeEvent(ei.Path(), s.Stream.EventType())
+			e, err := store.DecodeEvent(ei.Path(), s.Stream.EventType())
 			if err != nil {
 				s.errch <- err
 				break
@@ -185,23 +124,6 @@ func (s *Subscriber) watch(c chan notify.EventInfo) {
 	}
 }
 
-// decodeEvent is an internal helper that decodes a file at path and returns an
-// Event.
-func decodeEvent(path string, eventType reflect.Type) (Event, error) {
-	d := reflect.New(eventType)
-	b, err := ioutil.ReadFile(path)
-	if err != nil {
-		return Event{}, fmt.Errorf("error reading event data at %s: %s", path, err)
-	}
-	if err := json.Unmarshal(b, d.Interface()); err != nil {
-		return Event{}, fmt.Errorf("error unmarshaling event data from %s: %s", path, err)
-	}
-	return Event{
-		ID:   filepath.Base(path),
-		Data: d.Elem().Interface(),
-	}, nil
-}
-
 // NewSubscriberWithCallback is a helper that provides a very simple event loop
 // around a Subscriber. Events are passed to the callback function supplied by
 // cb.
@@ -210,7 +132,7 @@ func decodeEvent(path string, eventType reflect.Type) (Event, error) {
 // errors. The caller should ensure they are handling errors as normal by
 // waiting on the Done channel and processing any errors returned by the Error
 // function afterward.
-func NewSubscriberWithCallback(dir string, event interface{}, cb func(Event)) (*Subscriber, error) {
+func NewSubscriberWithCallback(dir string, event interface{}, cb func(store.Event)) (*Subscriber, error) {
 	s, err := NewSubscriber(dir, event)
 	if err != nil {
 		return nil, err
@@ -232,72 +154,4 @@ func NewSubscriberWithCallback(dir string, event interface{}, cb func(Event)) (*
 // is no longer needed. This performs a graceful shutdown of the subscriber.
 func (s *Subscriber) Close() {
 	close(s.errch)
-}
-
-// Dump dumps all of the events in the store for stream described by dir and
-// event. Technically, it's just dumping all of the events in the directory.
-// The events are returned as an Event slice.
-//
-// The order of the returned events is not deterministic. If it is up to the
-// consumer to structure the data or the handling of the data in a way that
-// facilitates proper hydration.
-func Dump(dir string, event interface{}) ([]Event, error) {
-	stream, err := store.NewStream(dir, event)
-	if err != nil {
-		return nil, err
-	}
-
-	return dump(stream)
-}
-
-func dump(s *store.Stream) ([]Event, error) {
-	entries, err := ioutil.ReadDir(s.Dir())
-	var es []Event
-	if err != nil {
-		return nil, fmt.Errorf("error reading event directory %s: %s", s.Dir(), err)
-	}
-	for _, f := range entries {
-		if !f.Mode().IsRegular() {
-			continue
-		}
-		e, err := decodeEvent(s.Dir()+"/"+f.Name(), s.EventType())
-		if err != nil {
-			return nil, err
-		}
-		es = append(es, e)
-	}
-	return es, nil
-}
-
-// DumpSorted works as per Dump, but sorts the returned events according to the
-// criteria defined by the event type's IndexedEvent interface. The function
-// will panic during sort if this interface is not implemented.
-func (s *Subscriber) DumpSorted(dir string, event interface{}) ([]Event, error) {
-	es, err := Dump(dir, event)
-	if err != nil {
-		return nil, err
-	}
-	sort.Sort(eventSlice(es))
-	return es, nil
-}
-
-// DumpSortedReverse acts as per DumpSorted, but reverses the sort order,
-// normally giving a descending order rather than an ascending one.
-func (s *Subscriber) DumpSortedReverse(dir string, event interface{}) ([]Event, error) {
-	es, err := Dump(dir, event)
-	if err != nil {
-		return nil, err
-	}
-	sort.Sort(sort.Reverse(eventSlice(es)))
-	return es, nil
-}
-
-// Fetch reads a single event from the store described by dir and type. The
-// supplied ID is essentially the file name.
-func (s *Subscriber) Fetch(dir string, event interface{}, id string) (Event, error) {
-	stream, err := store.NewStream(dir, event)
-	if err != nil {
-		return Event{}, err
-	}
-	return decodeEvent(stream.Dir()+"/"+id, stream.EventType())
 }
